Support persistent USB disable
Add a software option to disable USB for increased resiliency in
situations where Blinkekatzen are left unattended.
This feature is not invasive and does on purpose not touch any fuses
within the ESP. It may thus be circumventable by a skilled attacker.
Currently no more heavy-handed approaches like secure boot and fusing are
implemented to minimize the chance of bricking devices by accident during
development.