platform: improve robustness of platform detection
Introduce retry cycle to platform detection when detection of an
individual platform fails. This ensures sporadic EMI issues during
device startup or reboot do not cause the platform to be detected
incorrectly. Store name of last detected platform in NVS so that
platform can be probed first on next bootup to keep bootup times on all
supported platform low.